--- src/widgets/Sidebar/model/items.ts
@@ -1,38 +1,0 @@
-import React from 'react';
-import { RoutePath } from 'shared/config/routeConfig/routeConfig';
-import MainIcon from 'shared/assets/icons/Home.svg';
-import AbotIcon from 'shared/assets/icons/About.svg';
-import ProfileIcon from 'shared/assets/icons/Profile.svg';
-import ArticlesIcon from 'shared/assets/icons/Articles.svg';
-
-export interface SidebarItemType {
-    path: string;
-    text: string;
-    Icon: React.VFC<React.SVGProps<SVGSVGElement>>;
-    authOnly?: boolean;
-}
-
-export const SidebarItemList: SidebarItemType[] = [
-    {
-        path: RoutePath.main,
-        Icon: MainIcon,
-        text: 'Главная',
-    },
-    {
-        path: RoutePath.about,
-        Icon: AbotIcon,
-        text: 'О сайте',
-    },
-    {
-        path: `${RoutePath.profile}`,
-        Icon: ProfileIcon,
-        text: 'Профиль',
-        authOnly: true,
-    },
-    {
-        path: RoutePath.articles,
-        Icon: ArticlesIcon,
-        text: 'Статьи',
-        authOnly: true,
-    },
-];
